requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
-e git+https://github.com/city-of-helsinki/django-helusers.git#egg=django-helusers
-e git+https://github.com/juyrjola/django-orghierarchy.git#egg=django-orghierarchy
certifi==2018.10.15       # via requests
chardet==3.0.3            # via requests
defusedxml==0.5.0         # via python3-openid, social-auth-core
django-environ==0.4.5
django-extensions==2.1.3
django-filter==2.0.0
django-js-asset==1.1.0    # via django-mptt
django-markdownx==2.0.27
django-mptt==0.9.1
django-ordered-model==3.1.1
django==2.1.3
djangorestframework==3.9.0
drf-oidc-auth==0.9
future==0.17.1            # via pyjwkest
idna==2.7                 # via requests
markdown==3.0.1
oauthlib==2.1.0           # via requests-oauthlib, social-auth-core
pillow==5.3.0             # via django-markdownx
psycopg2-binary==2.7.6.1
pycryptodomex==3.7.0      # via pyjwkest
pyjwkest==1.4.0           # via drf-oidc-auth
pyjwt==1.6.4              # via social-auth-core
python3-openid==3.1.0     # via social-auth-core
pytz==2018.7              # via django
requests-oauthlib==1.0.0  # via social-auth-core
requests==2.20.1          # via pyjwkest, requests-oauthlib, social-auth-core
six==1.11.0               # via django-extensions, pyjwkest, social-auth-app-django, social-auth-core
social-auth-app-django==3.1.0
social-auth-core==2.0.0   # via social-auth-app-django
swapper==1.1.0
urllib3==1.24.1           # via requests
whitenoise==4.1.1
